+++ openfisca_core/commons/misc.py
@@ -1,14 +1,13 @@
-from typing import Optional, TypeVar
+from __future__ import annotations
 
+import numexpr
 import numpy
 
 from openfisca_core import types as t
 
-T = TypeVar("T")
 
-
-def empty_clone(original: T) -> T:
-    """Creates an empty instance of the same class of the original object.
+def empty_clone(original: object) -> object:
+    """Create an empty instance of the same class of the original object.
 
     Args:
         original: An object to clone.
@@ -30,13 +29,11 @@ def empty_clone(original: T) -> T:
         True
 
     """
-    Dummy: object
-    new: T
 
     Dummy = type(
         "Dummy",
         (original.__class__,),
-        {"__init__": lambda self: None},
+        {"__init__": lambda _: None},
     )
 
     new = Dummy()
@@ -44,8 +41,8 @@ def empty_clone(original: T) -> T:
     return new
 
 
-def stringify_array(array: Optional[t.Array[numpy.generic]]) -> str:
-    """Generates a clean string representation of a numpy array.
+def stringify_array(array: None | t.Array[numpy.generic]) -> str:
+    """Generate a clean string representation of a numpy array.
 
     Args:
         array: An array.
@@ -76,3 +73,33 @@ def stringify_array(array: Optional[t.Array[numpy.generic]]) -> str:
         return "None"
 
     return f"[{', '.join(str(cell) for cell in array)}]"
+
+
+def eval_expression(
+    expression: str,
+) -> str | t.Array[numpy.bool_] | t.Array[numpy.int32] | t.Array[numpy.float32]:
+    """Evaluate a string expression to a numpy array.
+
+    Args:
+        expression(str): An expression to evaluate.
+
+    Returns:
+        :obj:`object`: The result of the evaluation.
+
+    Examples:
+        >>> eval_expression("1 + 2")
+        array(3, dtype=int32)
+
+        >>> eval_expression("salary")
+        'salary'
+
+    """
+
+    try:
+        return numexpr.evaluate(expression)
+
+    except (KeyError, TypeError):
+        return expression
+
+
+__all__ = ["empty_clone", "eval_expression", "stringify_array"]