+++ openfisca_core/indexed_enums/__init__.py
@@ -21,6 +21,14 @@
 #
 # See: https://www.python.org/dev/peps/pep-0008/#imports
 
-from .config import ENUM_ARRAY_DTYPE  # noqa: F401
-from .enum import Enum  # noqa: F401
-from .enum_array import EnumArray  # noqa: F401
+from . import types
+from .config import ENUM_ARRAY_DTYPE
+from .enum import Enum
+from .enum_array import EnumArray
+
+__all__ = [
+    "ENUM_ARRAY_DTYPE",
+    "Enum",
+    "EnumArray",
+    "types",
+]