+++ openfisca_core/types.py
@@ -2,47 +2,57 @@ from __future__ import annotations
 
 from collections.abc import Iterable, Sequence, Sized
 from numpy.typing import NDArray
-from typing import Any, TypeVar, Union
-from typing_extensions import NewType, Protocol, TypeAlias
+from typing import Any, NewType, TypeVar, Union
+from typing_extensions import Protocol, TypeAlias
 
 import numpy
 import pendulum
 
-_N = TypeVar("_N", bound=numpy.generic, covariant=True)
+_N_co = TypeVar("_N_co", bound=numpy.generic, covariant=True)
 
 #: Type representing an numpy array.
-Array: TypeAlias = NDArray[_N]
+Array: TypeAlias = NDArray[_N_co]
 
 _L = TypeVar("_L")
 
 #: Type representing an array-like object.
 ArrayLike: TypeAlias = Sequence[_L]
 
-#: Type variable representing an error.
-_E = TypeVar("_E", covariant=True)
-
-#: Type variable representing a value.
-_A = TypeVar("_A", covariant=True)
-
 #: Generic type vars.
-_T_cov = TypeVar("_T_cov", covariant=True)
-_T_con = TypeVar("_T_con", contravariant=True)
+_T_co = TypeVar("_T_co", covariant=True)
 
 
 # Entities
 
+#: For example "person".
+EntityKey = NewType("EntityKey", str)
+
+#: For example "persons".
+EntityPlural = NewType("EntityPlural", str)
+
+#: For example "principal".
+RoleKey = NewType("RoleKey", str)
+
+#: For example "parents".
+RolePlural = NewType("RolePlural", str)
+
 
 class CoreEntity(Protocol):
-    key: Any
-    plural: Any
+    key: EntityKey
+    plural: EntityPlural
 
-    def check_role_validity(self, role: Any) -> None: ...
-    def check_variable_defined_for_entity(self, variable_name: Any) -> None: ...
+    def check_role_validity(self, role: object, /) -> None: ...
+    def check_variable_defined_for_entity(
+        self,
+        variable_name: VariableName,
+        /,
+    ) -> None: ...
     def get_variable(
         self,
-        variable_name: Any,
-        check_existence: Any = ...,
-    ) -> Any | None: ...
+        variable_name: VariableName,
+        check_existence: bool = ...,
+        /,
+    ) -> None | Variable: ...
 
 
 class SingleEntity(CoreEntity, Protocol): ...
@@ -52,20 +62,22 @@ class GroupEntity(CoreEntity, Protocol): ...
 
 
 class Role(Protocol):
-    entity: Any
+    entity: GroupEntity
     max: int | None
-    subroles: Any
+    subroles: None | Iterable[Role]
 
     @property
-    def key(self) -> str: ...
+    def key(self, /) -> RoleKey: ...
+    @property
+    def plural(self, /) -> None | RolePlural: ...
 
 
 # Holders
 
 
 class Holder(Protocol):
-    def clone(self, population: Any) -> Holder: ...
-    def get_memory_usage(self) -> Any: ...
+    def clone(self, population: Any, /) -> Holder: ...
+    def get_memory_usage(self, /) -> Any: ...
 
 
 # Parameters
@@ -83,13 +95,16 @@ InstantStr = NewType("InstantStr", str)
 PeriodStr = NewType("PeriodStr", str)
 
 
-class Indexable(Protocol[_T_cov]):
-    def __getitem__(self, index: int, /) -> _T_cov: ...
+class Container(Protocol[_T_co]):
+    def __contains__(self, item: object, /) -> bool: ...
 
 
-class DateUnit(Protocol):
-    def __contains__(self, other: object, /) -> bool: ...
-    def upper(self) -> str: ...
+class Indexable(Protocol[_T_co]):
+    def __getitem__(self, index: int, /) -> _T_co: ...
+
+
+class DateUnit(Container[str], Protocol):
+    def upper(self, /) -> str: ...
 
 
 class Instant(Indexable[int], Iterable[int], Sized, Protocol):
@@ -108,14 +123,14 @@ class Instant(Indexable[int], Iterable[int], Sized, Protocol):
 
 class Period(Indexable[Union[DateUnit, Instant, int]], Protocol):
     @property
-    def unit(self) -> DateUnit: ...
+    def unit(self, /) -> DateUnit: ...
     @property
-    def start(self) -> Instant: ...
+    def start(self, /) -> Instant: ...
     @property
-    def size(self) -> int: ...
+    def size(self, /) -> int: ...
     @property
-    def stop(self) -> Instant: ...
-    def offset(self, offset: str | int, unit: None | DateUnit = None) -> Period: ...
+    def stop(self, /) -> Instant: ...
+    def offset(self, offset: str | int, unit: None | DateUnit = None, /) -> Period: ...
 
 
 # Populations
@@ -124,17 +139,17 @@ class Period(Indexable[Union[DateUnit, Instant, int]], Protocol):
 class Population(Protocol):
     entity: Any
 
-    def get_holder(self, variable_name: Any) -> Any: ...
+    def get_holder(self, variable_name: VariableName, /) -> Any: ...
 
 
 # Simulations
 
 
 class Simulation(Protocol):
-    def calculate(self, variable_name: Any, period: Any) -> Any: ...
-    def calculate_add(self, variable_name: Any, period: Any) -> Any: ...
-    def calculate_divide(self, variable_name: Any, period: Any) -> Any: ...
-    def get_population(self, plural: Any | None) -> Any: ...
+    def calculate(self, variable_name: VariableName, period: Any, /) -> Any: ...
+    def calculate_add(self, variable_name: VariableName, period: Any, /) -> Any: ...
+    def calculate_divide(self, variable_name: VariableName, period: Any, /) -> Any: ...
+    def get_population(self, plural: None | str, /) -> Any: ...
 
 
 # Tax-Benefit systems
@@ -145,16 +160,21 @@ class TaxBenefitSystem(Protocol):
 
     def get_variable(
         self,
-        variable_name: Any,
-        check_existence: Any = ...,
-    ) -> Any | None: ...
+        variable_name: VariableName,
+        check_existence: bool = ...,
+        /,
+    ) -> None | Variable: ...
 
 
 # Variables
 
+#: For example "salary".
+VariableName = NewType("VariableName", str)
+
 
 class Variable(Protocol):
     entity: Any
+    name: VariableName
 
 
 class Formula(Protocol):
@@ -163,8 +183,9 @@ class Formula(Protocol):
         population: Population,
         instant: Instant,
         params: Params,
+        /,
     ) -> Array[Any]: ...
 
 
 class Params(Protocol):
-    def __call__(self, instant: Instant) -> ParameterNodeAtInstant: ...
+    def __call__(self, instant: Instant, /) -> ParameterNodeAtInstant: ...