+++ openfisca_core/variables/variable.py
@@ -9,7 +9,7 @@ import textwrap
 import numpy
 import sortedcontainers
 
-from openfisca_core import periods, tools, types as t
+from openfisca_core import commons, periods, types as t
 from openfisca_core.entities import Entity, GroupEntity
 from openfisca_core.indexed_enums import Enum, EnumArray
 from openfisca_core.periods import DateUnit, Period
@@ -450,7 +450,7 @@ class Variable:
                 )
         if self.value_type in (float, int) and isinstance(value, str):
             try:
-                value = tools.eval_expression(value)
+                value = commons.eval_expression(value)
             except SyntaxError:
                 msg = f"I couldn't understand '{value}' as a value for '{self.name}'"
                 raise ValueError(